include/util/object.h: attempt at implementing event consumation. EventTable is now merged into Object
 - event table class/header is now gone
 - event queue managed by the Engine... seems janky but it makes the most sense
     - the event table is also destroyed by Engine... still janky!
 - global enum EventTopic created to be able to isolate a specific reason for firing an event
 - global enum EventTarget made for event consumation
     - an object determines its EventTarget via a new virtual method get_target; any object which subscribes to events must implement it and return a UNIQUE EventTarget value
     - this value is used to do a cast of the "this" pointer in Object::consume so the member function pointers work properly

+++ src/lse/logger.cpp
@@ -95,7 +95,7 @@ void Logger::write_log(LogLevel log_level, std::ostream &stream, char *fmt_log) 
 
             if(location != nullptr) {
 
-                for(int i = 0; i < LOG_LINE_DELIM_LENGTH; ++i)
+                for(size_t i = 0; i < LOG_LINE_DELIM_LENGTH; ++i)
                     location[i] = '\0';
 
                 location += LOG_LINE_DELIM_LENGTH;
